fix: memberservice interface로 변경
